tools/map_compiler/requirements.txt
pydantic>=2.6.0
pymupdf>=1.23.0
Pillow>=10.0.0
google-genai>=0.3.0
